requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
sqlmodel==0.0.37
python-jose[cryptography]==3.3.0
pydantic-settings==2.7.1
email-validator==2.2.0
python-multipart==0.0.20
bcrypt==4.2.1
redis==5.2.1
celery==5.4.0
PyMySQL==1.1.1
alembic==1.14.0
openpyxl==3.1.5
pytest==8.3.4
httpx==0.28.1
